Update rabs_rcfp.f90
In my case, the rcfp_get_term_number returns fail=.true. even though the input is valid. This stops the whole program. The "fail" boolean variable is not initialized to .false. at the beginning of the function, thus, the function may return .true. even when the "value" is properly set in one of the "ifs". This issue may be system/compiler-dependent, depending on the internal initialization of the variable. A simple solution is proposed to add fail = .false. initialization on line 2345.